Switch VenueDetailsRow to VenueDetails
