--- www/src/js/views/venues/VenueDetailRow.scss
@@ -1,16 +1,0 @@
-@import "~styles/utils/modules-entry.scss";
-
-.venueDetailRow {
-  padding-top: 1rem;
-  border-top: 1px var(--gray-lighter) solid;
-
-  &:first-child {
-    padding-top: 0.25rem;
-    border-top: 0;
-  }
-}
-
-.venueTimetable {
-  composes: scrollable from global;
-  margin-bottom: 1rem;
-}
