+++ www/src/js/views/venues/VenueDetails.scss
@@ -1,0 +1,18 @@
+@import "~styles/utils/modules-entry.scss";
+
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 1.4rem;
+
+  h1 {
+    margin: 0;
+    font-weight: bold;
+    font-size: 1.6rem;
+  }
+}
+
+.timetable {
+  composes: scrollable from global;
+}
